fix(replay-anonymizer): close the host policy bypasses and extract the URL rules
Round two of the QA review compiled the host check and found it accepts
thirteen hosts it exists to reject. I reproduced every one before changing
anything.

A trailing dot defeated the whole name list. `rsplit('.')` on `localhost.`
returns the empty label after the final dot, which matches no suffix, so
`http://localhost./i.png` and `http://metadata.internal./i.png` were collected.
A fully qualified name resolves exactly like the bare one.

IPv6 skipped nearly every address rule. The old branch checked only loopback,
private and link-local on a v4-mapped address, so `[::ffff:0.0.0.0]`,
`[::ffff:100.64.0.1]`, `[::127.0.0.1]`, the 6to4 form `[2002:7f00:1::]`, the
NAT64 form `[64:ff9b::7f00:1]` and multicast `[ff02::1]` all passed. Every
address now goes through one IPv4 predicate, whichever way it was written.

The IPv4 rules also missed multicast, reserved 240/4, benchmarking 198.18/15
and 192.0.0.0/24.

The rules move to url_policy.rs, because they have a second consumer. The fetch
lane has to apply the same host rule again at request time, on the resolved
address and on every redirect hop, since a name that resolves publicly now can
resolve privately later. It also has to canonicalize what it fetched to find its
ledger entry. Keeping the policy in the module that accumulates one message's
URLs would have meant reimplementing it there.

`policy` moves behind CloudFront's marker. It was in the unconditional list,
which the file's own rule forbids: `policy=thumb` and `policy=full` are
plausible image selectors on an ordinary host.

The image-scrub consumer now rejects a URL ref. Widening the ref regex let one
through, and the shard store indexes by (pseudoTeam, hash) with the prefix
dropped, so it would have filed URL-sourced bytes in the slot of a
content-addressed image.
