+++ rust/replay-anonymizer/src/lib.rs
@@ -62,6 +62,8 @@ pub mod url;
 #[doc(hidden)]
 pub mod url_collect;
 #[doc(hidden)]
+pub mod url_policy;
+#[doc(hidden)]
 pub mod value;
 
 pub use allow_lists::AllowLists;
@@ -80,6 +82,7 @@ pub use timings::{PhaseTimings, PhaseTimingsSnapshot};
 #[cfg(feature = "typed-parse")]
 pub use typed::{parse_scrubbed_event, parse_scrubbed_event_with_ctx};
 pub use url_collect::{CollectedUrl, UrlCollection};
+pub use url_policy::{canonicalize, is_public_host, CanonicalUrl};
 
 /// Shared helpers for the image-neutralization tests across modules.
 #[cfg(test)]
